fix: export mask options types
