+++ src/interfaces/index.ts
@@ -1,3 +1,4 @@
 export * from './dot-notation.interface';
 export * from './fn.interface';
+export * from './string-utils.interface';
 export * from './value.interface';
